refactor(avatar/badge): optimize CSS
