+++ src/components/avatar/themes/avatar.base.scss
@@ -1,5 +1,8 @@
 @use 'styles/common/component';
 @use 'styles/utilities' as *;
+@use './light/themes' as *;
+
+$theme: $material;
 
 :host {
     display: inline-flex;
@@ -17,6 +20,15 @@
     flex-shrink: 0;
     overflow: hidden;
     font-family: var(--ig-font-family);
+    color: var-get($theme, 'color');
+    background: var-get($theme, 'background');
+    width: var-get($theme, 'size');
+    height: var-get($theme, 'size');
+
+    [part='initials'] {
+        font-size: calc(var-get($theme, 'size') / 2);
+        line-height: calc(var-get($theme, 'size') / 2);
+    }
 }
 
 img {
@@ -34,4 +46,12 @@ img {
 
 :host([shape='square']) [part='base'] {
     border-radius: 0;
+}
+
+:host([shape='rounded']) [part='base'] {
+    border-radius: var-get($theme, 'border-radius');
+}
+
+:host([shape='circle']) [part='base'] {
+    border-radius: calc(var-get($theme, 'size') / 2);
 }
