--- src/components/avatar/themes/shared/avatar.fluent.scss
@@ -1,24 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-[part='base'] {
-    color: var-get($theme, 'color');
-    background: var-get($theme, 'background');
-    width: var-get($theme, 'size');
-    height: var-get($theme, 'size');
-
-    [part='initials'] {
-        font-size: calc(var-get($theme, 'size') / 2);
-        line-height: calc(var-get($theme, 'size') / 2);
-    }
-}
-
-:host([shape='rounded']) [part='base'] {
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([shape='circle']) [part='base'] {
-    border-radius: calc(var-get($theme, 'size') / 2);
-}
