+++ src/components/avatar/themes/themes.ts
@@ -1,5 +1,7 @@
 import { css } from 'lit';
 
+// Shared Styles
+import { styles as base } from './avatar.base.css.js';
 // Dark Overrides
 import { styles as bootstrapDark } from './dark/avatar.bootstrap.css.js';
 import { styles as fluentDark } from './dark/avatar.fluent.css.js';
@@ -10,40 +12,35 @@ import { styles as bootstrapLight } from './light/avatar.bootstrap.css.js';
 import { styles as fluentLight } from './light/avatar.fluent.css.js';
 import { styles as indigoLight } from './light/avatar.indigo.css.js';
 import { styles as materialLight } from './light/avatar.material.css.js';
-// Shared Styles
-import { styles as bootstrap } from './shared/avatar.bootstrap.css.js';
-import { styles as fluent } from './shared/avatar.fluent.css.js';
-import { styles as indigo } from './shared/avatar.indigo.css.js';
-import { styles as material } from './shared/avatar.material.css.js';
 import type { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${base} ${bootstrapLight}
   `,
   material: css`
-    ${material} ${materialLight}
+    ${base} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${base} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${base} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${base} ${bootstrapDark}
   `,
   material: css`
-    ${material} ${materialDark}
+    ${base} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${base} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${base} ${indigoDark}
   `,
 };
