--- src/components/badge/themes/shared/badge.fluent.scss
@@ -1,22 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-::slotted(igc-icon) {
-    color: var-get($theme, 'icon-color');
-}
-
-:host([variant='primary']) {
-    background: var-get($theme, 'background-color');
-    color: var-get($theme, 'text-color');
-}
-
-:host([shape='square']),
-:host([outlined][shape='square']) [part='base'] {
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([outlined]) [part='base'] {
-    box-shadow: inset 0 0 0 rem(2px) var-get($theme, 'border-color');
-}
